requirements.txt
python-dotenv
anthropic
google-genai
requests
